check if exists before delete return msg

+++ src/main/java/com/luv2code/springboot/demo/rest/EmployeeRestController.java
@@ -65,8 +65,16 @@ public class EmployeeRestController {
     }
 
     @DeleteMapping("/employees/{employeeId}")
-    public void deleteEmployee(@PathVariable int employeeId) {
+    public String deleteEmployee(@PathVariable int employeeId) {
+
+        Employee tempEmployee = employeeService.findById(employeeId);
+
+        if (tempEmployee == null) {
+            throw new RuntimeException("Employee id not found - " + employeeId);
+        }
 
         employeeService.deleteById(employeeId);
+
+        return "Deleted employee id - " +  employeeId;
     }
 }
